Blacksholes SLR templatization and adding mem2stream kernel
