+++ FPGA/Xilinx/Batched/blacksholes/blacksholes_ops/blacksholes_ops.h
@@ -2,7 +2,7 @@
 #pragma once
 
 #include <ops_seq_v2.h>
-#include "blacksholes_common.h"
-#include "blacksholes_ops_kernels.h"
+#include "../blacksholes_common.h"
+
 
 int bs_explicit1_ops(float* result, OPS_instance * ops_inst, GridParameter gridData, BlacksholesParameter computeParam);
